requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2==2.9.9 
python-decouple==3.8